home: post history table page component revamp start
